feat: Initial commit. Switch from Scrapbin to Meowseum.
While transitioning to Git, I replaced each of my dependency packages
with a clone of a Git repository. I included a .txt containing a
structured list of these dependencies and a shortcut to the location
where these dependencies will be stored. I changed the path to
dependency binaries in settings.py.

The Meowseum version always features four buttons in the desktop header
regardless of whether the user is logged out, and the general Scrapbin
version only features three buttons. So, this first commit modifies
header_desktop.css to shorten the search bar so it will fit in the
smaller space.

Since the beginning of the project, I've customized bootstrap.css
because some of its selectors are specific enough that they'd be very
difficult to override in a separate style sheet. It branches off from
Bootstrap v3.3.7. Here are the changes I've made to
boostrap_customized.css:

- I added an -xl- breakpoint at 1800px for desktop-sized devices:
  1920x1080, 1ppx screens with a 23" diagonal at 96dpi.
- I added the range of viewport width "lsph" for "landscape phones and
  phablets", 544px to 767px. This is used in Bootstrap 4. They made
  room for it by declaring it to be the new "sm" and the old "sm" to
  "md", but for backward compatibility with what I'd already written, I
  couldn't do that.
- I made <video>s have vertical-align: middle; like Bootstrap uses for
  images, so GIFs inserted the same way as images will have the same
  margins.