fix: Correct "bonded with" links on slide page
First, the link with a pet's name to its slide page used "." instead
of "..", leading to the wrong URL. Second, when a pet was bonded with
more than one pet, the raw HTML was being shown. This was because my
humanize_list() function, which renders a Python list as "A, B, and C",
outputs a string. I forgot that I had to use mark_safe() on a string
to get Django to show it as HTML.