feat: Switch to using only session storage for backend
I decided that it'd be simpler to use only session storage for site
settings that are available whether or not the user is logged in,
rather than update UserProfile for a registered user. So, I removed
this code from toggle_night_mode.py. This decision will save the effort
of having to code every such feature twice. Finding out whether session
storage is slower than managing the database, significantly enough to
make this worth it, shouldn't be an issue until the project is
complete.