feat: Implement toggling only night mode sitewide CSS
The user can now toggle night mode via AJAX, but this only affects
the settings menu icon and the sitewide CSS. It does not yet affect
the CSS for the specific page, and it doesn't affect any of the other
images with night/day mode versions. In order to implement this, I had
to remove all the previous versions' JavaScript related to night mode.