fix: Post function was not posting data to the server
The ajax.post() function was failing to send the data in its data
argument to the server when three arguments were passed to the
function until this commit, because I forgot to add the data to
the $.ajax() options object in this scenario.

This commit also adds comments to ajax.post() which explain the
if-statements for handling optional arguments by clearly stating
which arguments are passed in each scenario. I changed the lettering
system for libary functions, perhaps temporarily, because
ajax.setupXCSRFTokenHeader() is only being used by ajax.post(). Now it
is indicated as a subroutine. I had it as having its own letter before
because it will be needed again if I write another function as basic as
ajax.post().