feat: Move comment deletion to a separate view
I moved the code related to a moderator deleting a comment into a
separate view. I also fixed a bug in which a moderator was unable
to delete a comment, because I'd removed code related to using
JavaScript without AJAX to make an element other than a submit button
submit the form.

I added a .action-only class for wrapping any element in a submit
button without adding any of the default submit button styling, so
that the trash can icons by the comments can submit the form.