feat: Following now works with JavaScript disabled
I moved the part of the Follow button which, when clicked, follows
or unfollows to using .ajax-form instead of .ajax-btn. However,
because the Follow option in the dropdown still uses .ajax-btn,
and their functions expect different kinds of response data
because I haven't synchronized them yet, the Follow option is
temporarily broken.

As of this commit, all the forms which contain no fields, aside from
the submit button itself, have had all their code moved from the slide
page view to separate views.