feat: Make .ajax-btn and .ajax-form compatible
I modified the functions related to .ajax-btn and .ajax-form so they
both take the same form of server response, then modified all the views
on the site which use these classes. I also updated the AJAX library
documentation. Compared to the previous commit, the only effect is that
the follow option in the follow button dropdown is no longer broken.