feat: Add ability to edit adoption record fields
I started working on the ability to edit the fields for an upload after
it is submitted. Currently, this only works for fields within the
Adoption model. I replaced the Delete Upload button on the slide page
with an Edit button, and deletion will be added to the edit page later.