fix: Bugs related to displaying a pet record
First, I fixed a bug at the top level of get_pet_information_record()
on the slide page. When I introduced it, I was attempting to test for
the category of an upload by checking for "if upload.adoption", but I
forgot that nonexistence evaluating to False only applies in a Django
template.

Second, I fixed bugs in the slide page that I introduced when
I switched my format for the choices for multiple choice fields. Before,
the database value was the same string as the human-readable label for
use in the form. I realized that that while the label should always be
capitalized, the database value should only sometimes be capitalized
so that Django can know whether it should be capitalized when using it
in the middle of a sentence. Until now, I'd forgotten to change the
conditions in the to use the lowercase value.

The coat description merged field will now display the coat colors even
when the pattern field is left blank. The eye color field now always
capitalizes its value. Before this commit, for a lost or found cat with
white "socks" and no other coat information, the value of the coat
description field would mistakenly have a ". " at the beginning, as in
 ". Her/his paws are white like socks."