fix: The internal ID field is now semi-required
The internal ID field is now required if the "bonded with" field is
filled out.

Before this commit, because the "bonded with" field assumes the
relationship is mutual, an error would arise when a user marks a pet as
bonded with another pet without filling out the ID field. When the
uploader of the recipient of this new relationship goes back to edit
the upload, this is when the exception would occur. Specifically,
the problem occurred during edit_upload.py's
initialize_bonded_with_form() with regard to the lines

for record in upload.adoption.bonded_with.all():
    list_of_IDs = list_of_IDs + [record.internal_id]
bonded_with_IDs = ', '.join(list_of_IDs)

In this situation, record.internal_id would be None, and the exception
happens when the join() method is expecting all of the elements of the
list to be strings. The resulting exception was

TypeError: sequence item 0: expected str instance, NoneType found

and making the internal ID field sometimes required fixed this.