fix: Minor syntax error which broke date widgets
I listed all of my front-end dependencies with a short description of
what each does. I fixed a bug that was causing the site's scripted date
widgets not to work because I forgot a ">" on the end of a script tag.

I removed the files related to Internet Explorer 8 support, because
according to NetMarketShare, its usage has dwindled to 1%.