refactor: Querystrings no longer need to begin with '?'
The redirect function will now add the '?' onto the beginning of the
querystring in order to be more DRY. Now it doesn't have to be included
at the beginning of the querystring argument. As well, I renamed the
"GET_args" keyword argument to "query". Both of these changes bring my
extended version of urlencode() into line with the behavior of the
original function.