refactor: Switch away from template_variables
Previously, I tried using a technique of building the dictionary
of variables to be sent to the template during the processing of
the view, rather than copying the variables to the dictionary at
the end of the view. I have found that this very rarely makes the
view more readable, because each addition to the dictionary
requires its own line. It also means that functions within the view
are likely to have bad packaging. If they use the content of the
template_variables dictionary and are copied into another view, it
makes it harder to see the function's input and output. So, I am
switching back to the standard method.

This commit switches everything except for the gallery-related views.
For those, which invoke a function in common_view_functions.py, I was
only able to change over everything before the invokation.

In the galleries related to a user's profile, which are the uploads,
likes, and comments galleries, I also deleted code related to
processing a follow request. This is because I have since created a
separate view which carries out a follow or unfollow action, and the
header now accesses it with AJAX, but I forgot to remove the code when
I added AJAX to the site.