fix: Redirect was unable to process URLs with arguments
redirect() was ignoring positional arguments after the 'to' argument.
Due to a misunderstanding of Python syntax, I placed '*args, **kwargs'
at the end of the function. In actuality, in Python 3, *args goes
after the positional arguments and before any keyword arguments.
This is important when a function is invoking another function by
passing *args and **kwargs, but it has keyword arguments which the
invoked function lacks. So, I rewrote redirect(),
ajaxWholePageRedirect(), and urlencode() to reflect this.