fix: Remove leftover search bar settings code
A long while ago, I removed a JavaScript setting for whether the
search bar will be shown when the page loads on mobile devices.
First, it was set up to keep the search bar open after scrolling it
down across page loads, and I realized it would be more annoying to
have to close it after searching. Second, it was causing page load
flash effects. This commit just removes some code that did nothing
because it was in an if suite with a predicate which would never
be true.