fix: Night mode/day mode toggling wasn't working
When I changed the syntax for .ajax-btn and .ajax-form returning a
response, I forgot to update toggle_night_mode.py. As a result, the
settings menu option for toggling night mode wasn't updating properly.
Because my JavaScript learns what the back end setting for night mode
is by inspecting the settings menu option, this broke the rest of
the night mode implementation by making the CSS change, but not the
images. Night mode is working properly now.