feat: Play button in mobile header no longer flashes
I moved the "mobile GIFs are playing" settings variable to the back
end. When GIFs in the gallery weren't set to autoplay, the icon would
initially show up as a pause icon for a split second before the
JavaScript changed it to a play icon. Using back end variables
in the template instead allowed the pause/play icon to be displayed
correctly the first time.