feat: Change uploads to partly use template inheritance
The adoption_upload_main.html, lost_upload_main.html, and
found_upload_main.html templates now inherit from a
petinfo_upload_main.html supertemplate. I didn't use template
inheritance directly on the template for the page so that the main part
of the form can potentially be used in a modal in the future.