refactor: Reorder fields in adoption/lost/found
I reordered the fields in the models and ModelForms for the adoption,
lost, and found upload forms, so that they are all in the same order
in which they appear in the actual form. I changed half of the
verbose_name arguments so that they are the same as the form labels,
but some of them contained HTML, and others were for Boolean fields
that should actually be part of a group, so I left them alone.