refactor: Replace adoption Boolean checkboxes
At first, I wanted to use a separate BooleanField for each field
in a checkbox group in order to make the syntax for querying easier.
The grouping itself is usually only relevant to the form. When
querying, I don't care about which group the "yes/no" field appeared
within. However, now that I'm storing labels on the back end, I see
that I need to use a CheckboxSelectMultiple in order to store one label
for the checkbox group itself in addition to the label for each
checkbox. Only this will eventually allow only having to specify labels
on the back end and prevent copying and pasting custom widget code
entirely someday.

I replaced the "likes cats", "likes dogs", and "likes kids"
BooleanFields with a "would prefer a household without" checkbox group.
I removed the "kids down to age" numerical field because I decided that
was being overly specific. I also replaced the "has been", as in
"has been house trained", BooleanFields with a checkbox group.