fix: humanize_list() has exception for empty field
After adding a new multiselect field to the database, the field will
have '' instead of an empty list in string delimiters, '[]', which
the field would have if the field were in the database before the
record was created. I didn't account for this scenario when
programming the humanize_list() error, so it caused the page to have
a SyntaxError exception in this situation until now.