refactor: Lost and Found forms now use checkbox groups
I switched the Lost and Found forms away from BooleanField checkboxes
with different names and a value of "on" to checkbox groups with the
same name and different values. See the commit message for making the
same update to the Adoption form.

I left the JavaScript intact for the nested form for the fields
relevant to the collar, microchip, or tattoo serial number. I had to
modify it so that it uses pure JS instead of referencing data-
attributes, however. I decided not to maintain the JavaScript for
linking the "spayed or neutered" and "has a spay or neuter tattoo"
together such that one is (un)checked if the other is (un)checked
in a way that doesn't make logical sense. It doesn't make sense
to do this until there is also validation on the back end.