fix: Pages weren't showing if pet was spayed or neutered
The functions which merge the field for the pet's sex, the field for
whether the pet is spayed or neutered, and the field for whether the
pet has a spay or neuter tattoo was displaying "Female" instead of
"Female, spayed" when applicable. My logic error was capitalizing
"male" or "female" and then testing using lowercase instead of "Male"
or "Female".