fix: Change validation tests for dependent forms
In the views with two forms working together for convenience, I
tested with "form_a.is_valid() and form_b.is_valid()". I read that
this would cause short circuiting to hide errors in form_b when both
forms had errors in the user input at the same time. However, for
unknown reasons, I was still able to observe a "Required field" error
in one form and a "must not contain commas" error in the other form at
the same time. Regardless, I rewrote the predicates so that variables
like form_a_is_valid are assigned first, then those are used for the
test.