feat: Update urls.py syntax to new standard
I replaced the string references to view paths with callables, because
the string syntax is retired in Django 1.10 (currently on 1.9.2). To be
able to do this, I had to modify views/__init__.py to specify the files
to be imported from the views directory.

I added a urls.py to the Meowseum app directory and used include() in
the project urls.py. This was part of Django's documentation's
tutorial, but not my textbook's, and serves to to make the app more
modular.

While implementing this, I deleted views/forms and
moved from_device.py to the main views directory, because I hadn't been
using it for any of the newer form-only related views.