feat: Move importing all files code to function
I moved the code in views/__init__.py, which allows urls.py to import
all files within views, to a function in common_views.py. I also
modified the project directory's urls.py to remove statements which
turned out to be unnecessary because they should be in the app
directory's urls.py instead.