fix: Exception when submitting a adopt/lost/found form
Whens submitting a form for an upload to the Adoption, Lost, or Found
category, the following exception would occur:
ValidationError: ["'' value must be either None, True or False."]

This was caused by a NullBooleanField model field using a ChoiceField,
which stores an empty string, instead of a NullBooleanField form field.
The error didn't occur until I switched from SQLite3 to PostgreSQL,
though based on the documentation, I should've expected it to occur.