feat: Verbose model names, improve UniquenessValidator
I added verbose model names to all of the models which needed them.
The difference is visible in the admin module and in the "unique"
error message. I ran some tests and improved the UniquenessValidator's
default error message to be exactly like Django's default "unique"
error message. These changes were made together because I needed
the verbose model names as part of the tests.