feat: Replace SignupForm and LoginForm with ModelForms
Now these forms are more DRY, because I'm not repeating any of the
built-in error messages. I won't have to update the maximum length of
the fields, which I'm keeping the same as Django's default, if the
maximum length changes again in a future Django version.

Additionally, I updated the error messages on LoginForm. Before, the
"email_or_username" field returned "Wrong email or password." even if
the user provided a username instead of an email. I added a shortcut
to Django's authentication models file for the duration of development.