refactor: Move contact info error into a template
I noticed the contact information error message's links were regular
HTML relative paths instead of Django's DRY links. This is the error
message which the user sees if the user hasn't provided contact
information which third party shelters. I moved the error
message into a template, and consequentially CONTACT_INFO_ERROR no
longer had to be sent to the template in the view.

While doing this, I noticed that from_device.py had an import statement
which is no longer being used, so I removed it.