refactor: Switch "Services included" to ChoiceArrayField
In the ShelterForm, I replaced four boolean fields for what services
the shelter includes with the adoption fee with a single
ChoiceArrayField. I also made the verbose_names more like the form
labels. I was about to switch to using template label variables, with
the labels hosted on the back end, when I noticed I hadn't made the
ChoiceArray change.