refactor: Reorganize forms.js
I added forms.improveBuiltInWidgets() only because I was considering
hiding the default option for an empty string in the <select> dropdown.
However, I realized I wanted the user to be able to unselect an option
after making a choice, and the option appearing in the menu isn't very
annoying, so I decided to revert this commit after saving it.