fix: User contact form wasn't loading saved info
In a previous commit, I tried refactoring to avoid repeating
specifying required fields. Assigning a value to "form.instance"
doesn't raise an exception, and the value changed when I inspected
the variables before and after assigning, but it didn't work for
loading the form information from the form. According to the forms
API part of the Django documentation, a form's data can't be altered
after the construction of the form. This commit rewrites the code
another way, by assigning the keyword arguments of the form to
variables, but this turned out to be more verbose, so I'm undoing it in
the next commit.