fix: Errors caused by field name changes
The template uses two different AddressForms. I learned I
needed to add a prefix to each form so that the fields would have
different names in order for the browser to know which fields go
with which AddressForm, so I implemented that. Next, I changed the
field names referenced by the "same as" widget.